Updated Feature Selection NB
Fixed bug at the end where the NB generates the results for the saturated and true subset models. Also changed how the data is generated, so nothing is hardcoded. The behavior of the GA is very odd in that if there's even a little noise, it performs poorly, and the best model is nearly saturated. Have to figure this out.